requirements.txt
requests
python-dotenv
pandas
langchain
langchain-openai
openai
